pyproject.toml
[project]
name = "vga-tt"
version = "0.1.0"
description = "VGA bouncing text for TinyTapeout"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "apio>=0.9.0",
    "cocotb>=1.8.0",
    "pytest",
    "pillow>=10.0.0",
    "imageio>=2.30.0",
    "numpy>=1.24.0",
]
